requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy<2
python-multipart==0.0.6
python-socketio[client]==5.14.3
aiortc
websockets==12.0
torch>=2.0.0
torchvision>=0.15.0
python-dotenv==1.0.0
av
ultralytics>=8.0.0
pandas
tqdm
seaborn
pymongo==4.6.1